+++ Controller/Payment/Index.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Satispay\Satispay\Controller\Payment;
 
 class Index extends \Magento\Framework\App\Action\Action
@@ -9,7 +10,8 @@ class Index extends \Magento\Framework\App\Action\Action
         \Magento\Framework\App\Action\Context $context,
         \Magento\Checkout\Model\Session $checkoutSession,
         \Satispay\Satispay\Model\Method\Satispay $satispay
-    ) {
+    )
+    {
         parent::__construct($context);
         $this->checkoutSession = $checkoutSession;
     }
@@ -17,13 +19,13 @@ class Index extends \Magento\Framework\App\Action\Action
     public function execute()
     {
         $order = $this->checkoutSession->getLastRealOrder();
-        
+
         if ($order->getState() == $order::STATE_NEW) {
             $satispayPayment = \SatispayGBusiness\Payment::create([
                 "flow" => "MATCH_CODE",
                 "amount_unit" => $order->getGrandTotal() * 100,
                 "currency" => $order->getOrderCurrencyCode(),
-                'description' => '#'.$order->getIncrementId(),
+                "external_code" => $order->getIncrementId(),
                 "callback_url" => $this->_url->getUrl('satispay/callback/', [
                     "_query" => "payment_id={uuid}"
                 ]),