+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## unreleased
+- Replace description with external_code in payment creation
+
 ## 2.0.3
 - Send new order email only when payment is accepted
 - Cancel order if payment is canceled